requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
decorator==5.0.9
Django==3.2.6
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
factory-boy==3.1.0
Faker==8.11.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.1
packaging==21.0
pandas==1.3.1
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
rich==10.7.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
